refactor: stateless provider architecture with direct LangChain instance creation
- Providers are now completely stateless, each defined in a single file
- Added createChatInstance(auth, modelId, options) to provider interface
- Added createEmbeddingInstance(auth, modelId) for embedding providers
- Removed createRuntimeDefinition in favor of direct instance creation
- Simplified ProviderRegistry to just track configured providers + auth
- Agent.chooseModel() now creates fresh LangChain instances per model
- Removed base runtime files (openaiCompatible.ts, anthropicRuntime.ts, ollamaRuntime.ts)
- Cleaned up unused types (ChatModelFactory, EmbeddingModelFactory, RuntimeProviderDefinition)
