--- src/agent/providers/index.ts
@@ -1,29 +1,0 @@
-// Provider Registry
-export { ProviderRegistry } from "./ProviderRegistry";
-
-// Types
-export type {
-	ModelOptions,
-	ChatModelFactory,
-	EmbeddingModelFactory,
-	ProviderDefinition,
-	BuiltInProviderOptions,
-	BuiltInProviderModelMap,
-	BuiltInProviderModelMapEntry,
-	ProviderApiConfigOptions,
-} from "./types";
-
-// Errors
-export {
-	ProviderRegistryError,
-	ProviderNotFoundError,
-	ModelNotFoundError,
-	ProviderAuthError,
-	ProviderEndpointError,
-	ProviderImportError,
-} from "./errors";
-
-// Provider factories (for custom usage)
-export { createOpenAIProviderDefinition } from "./openai";
-export { createAnthropicProviderDefinition } from "./anthropic";
-export { createOllamaProviderDefinition } from "./ollama";
