+++ src/components/ui/logos/GenericAIIcon.svelte
@@ -2,15 +2,16 @@
 import { icon } from "../../../utils/utils";
 
 interface Props {
-	size?: { width?: number; height?: number };
-	className?: string;
+	width?: number;
+	height?: number;
+	class?: string;
 }
 
-const { size = { width: 16, height: 16 }, className = "" }: Props = $props();
+let { width = 16, height = 16, class: className = "" }: Props = $props();
 </script>
 
 <div
-    class="p-0 flex items-center {className}"
-    use:icon={"sparkles"}
-    style={"--icon-size: " + (size ? size : "var(--icon-s)")}
+	class="p-0 flex items-center {className}"
+	use:icon={"sparkles"}
+	style="--icon-size: {width}px"
 ></div>
