+++ src/components/ui/logos/OllamaLogo.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+interface Props {
+	width?: number;
+	height?: number;
+	class?: string;
+}
+
+let { width = 16, height = 16, class: className = "" }: Props = $props();
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	viewBox="0 0 1024 1024"
+	{width}
+	{height}
+	class={className}
+	fill="currentColor"
+>
+	<path d="M684.3 507.4c0-25.2-3.8-49.6-10.9-72.5l-.8-2.3c-.5-1.4-1-2.9-1.5-4.3-.1-.3-.2-.5-.3-.8-5.4-14.8-12.4-28.8-20.7-41.7-1.5-2.3-3-4.6-4.6-6.8l-.8-1.1c-.3-.5-.7-1-1-1.5-8.5-11.8-18.3-22.6-29.2-32.2l-.3-.3c-.9-.8-1.8-1.6-2.7-2.4-1.2-1-2.3-2-3.5-3-1.6-1.3-3.2-2.6-4.8-3.9-1-.8-2-1.5-3-2.3-.7-.5-1.5-1.1-2.2-1.6-1.9-1.4-3.9-2.7-5.9-4-5.7-3.8-11.7-7.3-17.9-10.5-.1 0-.2-.1-.2-.1-1.7-.9-3.4-1.7-5.1-2.5-.6-.3-1.2-.6-1.9-.9-1.1-.5-2.2-1-3.4-1.5-.6-.3-1.2-.5-1.8-.8-1.5-.6-2.9-1.2-4.4-1.8-.1 0-.2-.1-.3-.1-1.5-.6-2.9-1.1-4.4-1.7-.8-.3-1.7-.6-2.5-.9l-4.2-1.5c-.8-.3-1.6-.5-2.4-.8-1.5-.5-3-1-4.6-1.5-.6-.2-1.2-.3-1.8-.5l-5.4-1.5c-.5-.1-1-.3-1.6-.4-2-.5-4-1-6-1.4-.3-.1-.7-.1-1-.2-1.9-.4-3.7-.8-5.6-1.1-.7-.1-1.5-.3-2.2-.4-1.6-.3-3.1-.5-4.7-.8-.8-.1-1.6-.2-2.5-.3-1.6-.2-3.2-.4-4.8-.5-.8-.1-1.5-.2-2.3-.2-1.8-.2-3.6-.3-5.3-.4-.6 0-1.3-.1-1.9-.1-2.3-.1-4.7-.2-7-.3h-4c-2.4 0-4.7.1-7.1.3-.6 0-1.3.1-1.9.1-1.8.1-3.6.2-5.4.4-.8.1-1.5.1-2.3.2-1.6.2-3.2.3-4.8.5-.8.1-1.6.2-2.4.3-1.6.2-3.2.5-4.8.8-.7.1-1.5.3-2.2.4-1.9.4-3.8.7-5.6 1.1-.3.1-.7.1-1 .2-2 .5-4 .9-6 1.4-.5.1-1 .2-1.5.4l-5.4 1.5c-.6.2-1.2.4-1.8.5-1.5.5-3 .9-4.5 1.4-.8.3-1.6.5-2.4.8l-4.2 1.5c-.9.3-1.7.6-2.6.9-1.4.5-2.9 1.1-4.3 1.7l-.4.1c-1.5.6-2.9 1.2-4.4 1.8-.6.3-1.2.5-1.8.8-1.1.5-2.2 1-3.4 1.5-.6.3-1.2.6-1.9.9-1.7.8-3.4 1.7-5.1 2.5l-.2.1c-6.2 3.2-12.2 6.7-17.9 10.5-2 1.3-4 2.7-5.9 4-.8.5-1.5 1.1-2.2 1.6-1 .7-2 1.5-3 2.3-1.6 1.3-3.2 2.6-4.8 3.9-1.2 1-2.4 2-3.5 3-.9.8-1.8 1.6-2.7 2.4l-.3.3c-10.9 9.6-20.7 20.4-29.2 32.2-.3.5-.7.9-1 1.4l-.8 1.2c-1.6 2.3-3.1 4.6-4.6 6.9-8.3 12.9-15.3 26.9-20.7 41.7-.1.3-.2.5-.3.8-.5 1.4-1 2.9-1.5 4.3l-.8 2.3c-7.1 22.9-10.9 47.2-10.9 72.4 0 25.2 3.8 49.6 10.9 72.5l.8 2.3c.5 1.4 1 2.9 1.5 4.3.1.3.2.5.3.8 5.4 14.8 12.4 28.8 20.7 41.7 1.5 2.3 3 4.6 4.6 6.8l.8 1.2c.3.5.7.9 1 1.4 8.5 11.8 18.3 22.6 29.2 32.2l.3.3c.9.8 1.8 1.6 2.7 2.4 1.2 1 2.3 2 3.5 3 1.6 1.3 3.2 2.6 4.8 3.9 1 .8 2 1.5 3 2.3.7.5 1.5 1.1 2.2 1.6 1.9 1.4 3.9 2.7 5.9 4 5.7 3.8 11.7 7.3 17.9 10.5.1 0 .2.1.2.1 1.7.9 3.4 1.7 5.1 2.5.6.3 1.2.6 1.9.9 1.1.5 2.2 1 3.4 1.5.6.3 1.2.5 1.8.8 1.5.6 2.9 1.2 4.4 1.8.1 0 .2.1.4.1 1.4.6 2.9 1.1 4.3 1.7.9.3 1.7.6 2.6.9l4.2 1.5c.8.3 1.6.5 2.4.8 1.5.5 3 .9 4.5 1.4.6.2 1.2.4 1.8.5l5.4 1.5c.5.1 1 .3 1.5.4 2 .5 4 1 6 1.4.3.1.7.1 1 .2 1.9.4 3.7.8 5.6 1.1.7.1 1.5.3 2.2.4 1.6.3 3.2.5 4.8.8.8.1 1.6.2 2.4.3 1.6.2 3.2.4 4.8.5.8.1 1.5.2 2.3.2 1.8.2 3.6.3 5.4.4.6 0 1.3.1 1.9.1 2.4.1 4.7.3 7.1.3h3.9c2.3 0 4.7-.1 7-.3.6 0 1.3-.1 1.9-.1 1.8-.1 3.6-.2 5.3-.4.8-.1 1.5-.1 2.3-.2 1.6-.2 3.2-.3 4.8-.5.8-.1 1.6-.2 2.5-.3 1.6-.2 3.1-.5 4.7-.8.7-.1 1.5-.3 2.2-.4 1.9-.4 3.8-.7 5.6-1.1.3-.1.7-.1 1-.2 2-.5 4-.9 6-1.4.5-.1 1-.2 1.6-.4l5.4-1.5c.6-.2 1.2-.3 1.8-.5 1.5-.5 3-.9 4.6-1.5.8-.3 1.6-.5 2.4-.8l4.2-1.5c.8-.3 1.7-.6 2.5-.9 1.5-.5 2.9-1.1 4.4-1.7l.3-.1c1.5-.6 3-1.2 4.4-1.8.6-.3 1.2-.5 1.8-.8 1.1-.5 2.2-1 3.4-1.5.6-.3 1.3-.6 1.9-.9 1.7-.8 3.4-1.7 5.1-2.5.1 0 .2-.1.2-.1 6.2-3.2 12.2-6.7 17.9-10.5 2-1.3 4-2.7 5.9-4 .8-.5 1.5-1.1 2.2-1.6 1-.7 2-1.5 3-2.3 1.6-1.3 3.2-2.6 4.8-3.9 1.2-1 2.4-2 3.5-3 .9-.8 1.8-1.6 2.7-2.4l.3-.3c10.9-9.6 20.7-20.4 29.2-32.2.3-.5.7-1 1-1.5l.8-1.1c1.6-2.3 3.1-4.5 4.6-6.8 8.3-12.9 15.3-26.9 20.7-41.7.1-.3.2-.5.3-.8.5-1.4 1-2.9 1.5-4.3l.8-2.3c7.1-22.8 10.9-47.2 10.9-72.4zM512 678.8c-94.6 0-171.4-76.8-171.4-171.4 0-94.6 76.8-171.4 171.4-171.4 94.6 0 171.4 76.8 171.4 171.4 0 94.6-76.8 171.4-171.4 171.4z"/>
+	<circle cx="512" cy="507.4" r="114.3"/>
+	<ellipse cx="416.8" cy="264.4" rx="56.8" ry="68.5"/>
+	<ellipse cx="607.2" cy="264.4" rx="56.8" ry="68.5"/>
+</svg>
