--- src/providers/base/index.ts
@@ -1,22 +1,0 @@
-/**
- * Base runtime exports
- *
- * This module exports factory functions for creating LangChain models
- * for different provider runtimes (OpenAI-compatible, Anthropic, Ollama).
- */
-
-export {
-	createOpenAIChatModel,
-	createOpenAIEmbeddingModel,
-	discoverOpenAIModels,
-	validateOpenAIAuth,
-} from "./openaiCompatible";
-
-export { createAnthropicChatModel, validateAnthropicAuth } from "./anthropicRuntime";
-
-export {
-	createOllamaChatModel,
-	createOllamaEmbeddingModel,
-	discoverOllamaModels,
-	validateOllamaConnection,
-} from "./ollamaRuntime";
