+++ src/providers/builtin/openai.ts
@@ -9,37 +9,83 @@
  * Authentication: Field-based with apiKey (required), baseUrl (optional), headers (optional)
  */
 
-import {
-	createOpenAIChatModel,
-	createOpenAIEmbeddingModel,
-	discoverOpenAIModels,
-	validateOpenAIAuth,
-} from "../base/openaiCompatible";
+import { ChatOpenAI, OpenAIEmbeddings } from "@langchain/openai";
+import OpenAILogo from "../../components/ui/logos/OpenAILogo.svelte";
+import { ProviderAuthError, ProviderEndpointError } from "../errors";
 import { buildFieldBasedAuth } from "../helpers";
 import type {
 	AuthValidationResult,
 	BuiltInProviderDefinition,
 	ChatModelConfig,
-	ChatModelFactory,
 	DiscoveredModels,
-	EmbeddingModelFactory,
-	ModelOptions,
 	RuntimeAuthState,
 	RuntimeFieldBasedAuthState,
-	RuntimeProviderDefinition,
 } from "../types";
 
-/**
- * Default chat models available for OpenAI.
- * These are used to create the initial runtime definition.
- * Users can discover more models via model discovery.
- */
-const DEFAULT_CHAT_MODELS = ["gpt-4o", "gpt-4o-mini", "gpt-4-turbo", "gpt-3.5-turbo"];
+// =============================================================================
+// Constants
+// =============================================================================
+
+/** Default base URL for OpenAI API */
+const OPENAI_DEFAULT_BASE_URL = "https://api.openai.com/v1";
+
+// =============================================================================
+// Helper Functions
+// =============================================================================
 
 /**
- * Default embedding models available for OpenAI.
+ * Removes trailing slashes from a URL.
+ * @param url - The URL to sanitize
+ * @returns URL with trailing slashes removed
  */
-const DEFAULT_EMBEDDING_MODELS = ["text-embedding-3-small", "text-embedding-3-large", "text-embedding-ada-002"];
+function sanitizeBaseUrl(url: string): string {
+	return url.replace(/\/+$/, "");
+}
+
+/**
+ * Safely reads response text, returning undefined on error.
+ */
+async function safeReadText(response: Response): Promise<string | undefined> {
+	try {
+		return await response.text();
+	} catch {
+		return undefined;
+	}
+}
+
+/**
+ * Converts an error to a ProviderEndpointError.
+ */
+function toEndpointError(error: unknown): Error {
+	if (error instanceof ProviderEndpointError) {
+		return error;
+	}
+	const message = error instanceof Error ? error.message : String(error);
+	return new ProviderEndpointError("openai", message);
+}
+
+// =============================================================================
+// API Response Types
+// =============================================================================
+
+/**
+ * OpenAI API models response structure.
+ */
+interface OpenAIModelResponse {
+	data?: OpenAIModelResource[];
+}
+
+/**
+ * OpenAI API model resource structure.
+ */
+interface OpenAIModelResource {
+	id?: string;
+	object?: string;
+}
+
+// =============================================================================
+// Provider Definition
+// =============================================================================
 
 /**
  * OpenAI built-in provider definition.
@@ -55,6 +101,7 @@ export const openaiProvider: BuiltInProviderDefinition = {
 	// =========================================================================
 	id: "openai",
 	displayName: "OpenAI",
+	logo: OpenAILogo,
 	isBuiltIn: true,
 
 	// =========================================================================
@@ -95,49 +142,74 @@ export const openaiProvider: BuiltInProviderDefinition = {
 	// =========================================================================
 
 	/**
-	 * Creates the runtime provider definition with model factories.
+	 * Creates a LangChain ChatOpenAI instance.
 	 *
-	 * The factories capture the auth state in closures so models can be
-	 * created without re-passing credentials.
+	 * @param auth - Runtime authentication state with resolved secrets
+	 * @param modelId - The model ID (e.g., "gpt-4o", "gpt-4o-mini")
+	 * @param options - Optional model configuration (temperature, etc.)
+	 * @returns ChatOpenAI instance configured with auth and model
 	 */
-	createRuntimeDefinition: async (auth: RuntimeAuthState): Promise<RuntimeProviderDefinition> => {
+	createChatInstance: (auth: RuntimeAuthState, modelId: string, options?: Partial<ChatModelConfig>): ChatOpenAI => {
 		if (auth.type !== "field-based") {
 			throw new Error("OpenAI provider requires field-based authentication");
 		}
 
 		const fieldAuth = auth as RuntimeFieldBasedAuthState;
-
-		// Create chat model factories
-		const chatModels: Record<string, ChatModelFactory> = {};
-		for (const modelId of DEFAULT_CHAT_MODELS) {
-			chatModels[modelId] = async (options?: ModelOptions) => {
-				// Cast ModelOptions to ChatModelConfig if provided
-				const chatConfig = options as ChatModelConfig | undefined;
-				return createOpenAIChatModel(modelId, fieldAuth, chatConfig);
-			};
-		}
-
-		// Create embedding model factories
-		const embeddingModels: Record<string, EmbeddingModelFactory> = {};
-		for (const modelId of DEFAULT_EMBEDDING_MODELS) {
-			embeddingModels[modelId] = async () => {
-				return createOpenAIEmbeddingModel(modelId, fieldAuth);
-			};
-		}
-
-		return {
-			chatModels,
-			embeddingModels,
-			defaultChatModel: "gpt-4o",
-			defaultEmbeddingModel: "text-embedding-3-small",
+		const config: Record<string, unknown> = {
+			model: modelId,
+			apiKey: fieldAuth.values.apiKey,
 		};
+
+		// Add temperature if provided
+		if (options?.temperature !== undefined) {
+			config.temperature = options.temperature;
+		}
+
+		// Add baseUrl configuration if provided and not empty
+		const baseUrl = fieldAuth.values.baseUrl;
+		if (baseUrl && baseUrl.trim() !== "") {
+			config.configuration = {
+				baseURL: sanitizeBaseUrl(baseUrl),
+			};
+		}
+
+		return new ChatOpenAI(config);
+	},
+
+	/**
+	 * Creates a LangChain OpenAIEmbeddings instance.
+	 *
+	 * @param auth - Runtime authentication state with resolved secrets
+	 * @param modelId - The model ID (e.g., "text-embedding-3-small")
+	 * @returns OpenAIEmbeddings instance configured with auth and model
+	 */
+	createEmbeddingInstance: (auth: RuntimeAuthState, modelId: string): OpenAIEmbeddings => {
+		if (auth.type !== "field-based") {
+			throw new Error("OpenAI provider requires field-based authentication");
+		}
+
+		const fieldAuth = auth as RuntimeFieldBasedAuthState;
+		const config: Record<string, unknown> = {
+			model: modelId,
+			apiKey: fieldAuth.values.apiKey,
+		};
+
+		// Add baseUrl configuration if provided and not empty
+		const baseUrl = fieldAuth.values.baseUrl;
+		if (baseUrl && baseUrl.trim() !== "") {
+			config.configuration = {
+				baseURL: sanitizeBaseUrl(baseUrl),
+			};
+		}
+
+		return new OpenAIEmbeddings(config);
 	},
 
 	/**
 	 * Validates authentication credentials.
 	 *
-	 * First checks that required fields are present, then delegates to
-	 * the base runtime's validateOpenAIAuth for API-level validation.
+	 * Makes a GET request to the /models endpoint to verify the API key is valid.
+	 * This validates both the API key and the baseUrl configuration.
 	 */
 	validateAuth: async (auth: RuntimeAuthState): Promise<AuthValidationResult> => {
 		// Check auth type
@@ -153,20 +225,131 @@ export const openaiProvider: BuiltInProviderDefinition = {
 			return { valid: false, error: "API key is required" };
 		}
 
-		// Delegate to base runtime for API validation
-		return validateOpenAIAuth(fieldAuth);
+		const baseUrl =
+			fieldAuth.values.baseUrl && fieldAuth.values.baseUrl.trim() !== ""
+				? sanitizeBaseUrl(fieldAuth.values.baseUrl)
+				: OPENAI_DEFAULT_BASE_URL;
+
+		let response: Response;
+		try {
+			response = await globalThis.fetch(`${baseUrl}/models`, {
+				method: "GET",
+				headers: {
+					Authorization: `Bearer ${apiKey}`,
+					"Content-Type": "application/json",
+				},
+			});
+		} catch (error) {
+			const message = error instanceof Error ? error.message : String(error);
+			return { valid: false, error: `Connection failed: ${message}` };
+		}
+
+		if (response.ok) {
+			return { valid: true };
+		}
+
+		// Handle error response
+		const errorBody = await safeReadText(response);
+		let errorCode: string | undefined;
+		let errorMessage: string | undefined;
+		try {
+			const parsed = errorBody
+				? (JSON.parse(errorBody) as {
+						error?: { code?: string; message?: string };
+					})
+				: undefined;
+			errorCode = parsed?.error?.code;
+			errorMessage = parsed?.error?.message;
+		} catch {
+			// ignore parse errors
+		}
+
+		// Build error message
+		if (response.status === 401 || response.status === 403 || errorCode === "invalid_api_key") {
+			const detail = errorMessage || `Authentication failed (${response.status})`;
+			return { valid: false, error: detail };
+		}
+
+		// Other errors
+		const detail = errorMessage || errorBody || `Request failed with status ${response.status}`;
+		return { valid: false, error: detail };
 	},
 
 	/**
 	 * Discovers available models from OpenAI API.
 	 *
-	 * Delegates to the base runtime's discoverOpenAIModels function.
+	 * Makes a GET request to /models endpoint and returns all model IDs.
+	 * Models are returned in both chat and embedding arrays - users choose
+	 * which model to use for each purpose.
 	 */
 	discoverModels: async (auth: RuntimeAuthState): Promise<DiscoveredModels> => {
 		if (auth.type !== "field-based") {
 			throw new Error("OpenAI provider requires field-based authentication");
 		}
 
-		return discoverOpenAIModels(auth as RuntimeFieldBasedAuthState);
+		const fieldAuth = auth as RuntimeFieldBasedAuthState;
+		const apiKey = fieldAuth.values.apiKey;
+		if (!apiKey) {
+			throw new Error("OpenAI model discovery requires an API key.");
+		}
+
+		const baseUrl =
+			fieldAuth.values.baseUrl && fieldAuth.values.baseUrl.trim() !== ""
+				? sanitizeBaseUrl(fieldAuth.values.baseUrl)
+				: OPENAI_DEFAULT_BASE_URL;
+
+		let response: Response;
+		try {
+			response = await globalThis.fetch(`${baseUrl}/models`, {
+				method: "GET",
+				headers: {
+					Authorization: `Bearer ${apiKey}`,
+					"Content-Type": "application/json",
+				},
+			});
+		} catch (error) {
+			throw toEndpointError(error);
+		}
+
+		if (!response.ok) {
+			const errorBody = await safeReadText(response);
+			let errorCode: string | undefined;
+			let errorMessage: string | undefined;
+			try {
+				const parsed = errorBody
+					? (JSON.parse(errorBody) as {
+							error?: { code?: string; message?: string };
+						})
+					: undefined;
+				errorCode = parsed?.error?.code;
+				errorMessage = parsed?.error?.message;
+			} catch {
+				// ignore parse errors
+			}
+
+			if (response.status === 401 || response.status === 403 || errorCode === "invalid_api_key") {
+				throw new ProviderAuthError("openai", response.status, errorCode, errorMessage);
+			}
+			throw new Error(
+				`OpenAI model discovery failed with status ${response.status}${errorBody ? `: ${errorBody}` : ""}`,
+			);
+		}
+
+		const payload = (await response.json()) as OpenAIModelResponse;
+		const resources = Array.isArray(payload.data) ? payload.data : [];
+
+		const models: string[] = [];
+
+		for (const resource of resources) {
+			const id = typeof resource?.id === "string" ? resource.id.trim() : undefined;
+			if (!id) {
+				continue;
+			}
+
+			models.push(id);
+		}
+
+		// Return all models in both arrays - let users choose which model to use for each purpose
+		return { chat: models, embedding: models };
 	},
 };
