+++ src/types/provider/runtime.ts
@@ -52,46 +52,3 @@ export interface RuntimeOAuthAuthState {
  * Unlike StoredAuthState, all secrets have been resolved from SecretStorage.
  */
 export type RuntimeAuthState = RuntimeFieldBasedAuthState | RuntimeOAuthAuthState;
-
-/**
- * Options passed to model factory functions.
- */
-export type ModelOptions = Record<string, unknown>;
-
-/**
- * Factory function that creates a LangChain chat model instance.
- */
-export type ChatModelFactory = (options?: ModelOptions) => Promise<unknown>;
-
-/**
- * Factory function that creates a LangChain embedding model instance.
- */
-export type EmbeddingModelFactory = (options?: ModelOptions) => Promise<unknown>;
-
-/**
- * Runtime provider definition containing model factories.
- *
- * Created by a provider's `createRuntimeDefinition` method
- * and used by the ProviderRegistry to instantiate models at runtime.
- */
-export interface RuntimeProviderDefinition {
-	/**
-	 * Map of model name → factory function for chat models.
-	 */
-	chatModels: Record<string, ChatModelFactory>;
-
-	/**
-	 * Map of model name → factory function for embedding models.
-	 */
-	embeddingModels: Record<string, EmbeddingModelFactory>;
-
-	/**
-	 * Default chat model to use if none specified.
-	 */
-	defaultChatModel?: string;
-
-	/**
-	 * Default embedding model to use if none specified.
-	 */
-	defaultEmbeddingModel?: string;
-}
